[CI] Test C API demo (#6159)
* Fix CMake install config to use dependencies

* [CI] Test C API demo

* Explicitly cast num_feature, to avoid warning in Linux

+++ demo/c-api/c-api-demo.c
@@ -62,7 +62,7 @@ int main(int argc, char** argv) {
 
   bst_ulong num_feature = 0;
   safe_xgboost(XGBoosterGetNumFeature(booster, &num_feature));
-  printf("num_feature: %llu\n", num_feature);
+  printf("num_feature: %lu\n", (unsigned long)(num_feature));
 
   // predict
   bst_ulong out_len = 0;
